Weakened equivalence construction by size
It is not necessary for two type spaces to have the same size to
construct an equivalence. As long the size is the same for the things
that are equivalent, it is OK. This also makes equivalences different
from isomorphisms, since isomorphisms requires type spaces to have the
same size.